feat: redo daily components
