+++ src/common/locales/zh.json
@@ -703,8 +703,8 @@
     },
     "charts": {
       "tokenUsageTrend": "令牌使用趋势",
-      "dailyCostBreakdown": "每日成本分解",
-      "dailyMessageBreakdown": "每日消息分解",
+      "costBreakdown": "",
+      "dMessageBreakdown": "",
       "modelUsageDistribution": "模型使用分布",
       "inputTokens": "输入令牌",
       "outputTokens": "输出令牌",
