+++ src/renderer/src/components/usage/utils.ts
@@ -26,3 +26,44 @@ export const generateColorPalette = (count: number): string[] => {
 
   return colors;
 };
+
+enum GroupBy {
+  Year = 'year',
+  Month = 'month',
+  Day = 'day',
+  Hour = 'hour',
+}
+
+export const getPeriodKey = (timestamp: string, groupBy: GroupBy): string => {
+  const dateObj = new Date(timestamp);
+
+  switch (groupBy) {
+    case GroupBy.Hour:
+      return dateObj.toISOString().slice(0, 13); // "YYYY-MM-DDTHH"
+    case GroupBy.Day:
+      return dateObj.toISOString().split('T')[0]; // "YYYY-MM-DD"
+    case GroupBy.Year:
+      return dateObj.toISOString().slice(0, 4); // "YYYY"
+    // default is month
+    default:
+      return dateObj.toISOString().slice(0, 7); // "YYYY-MM"
+  }
+};
+
+export const formatDateByGroup = (date: Date, groupBy: GroupBy): string => {
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1; // Months are 0-indexed
+
+  switch (groupBy) {
+    case GroupBy.Year:
+      return year.toString();
+    case GroupBy.Month:
+      return `${month}/${year}`;
+    case GroupBy.Day:
+      return date.toLocaleDateString();
+    case GroupBy.Hour:
+      return date.toLocaleTimeString([], { hour: '2-digit' });
+    default:
+      return date.toLocaleDateString();
+  }
+};
